fix bug with includes of undefined

+++ packages/rnv/plugins/@react-native-community/overrides/cli/build/commands/runIOS/runIOS.js
@@ -241,7 +241,7 @@ async function runOnDevice(selectedDevice, scheme, xcodeProject, configuration, 
     // check if brew is used and if it is, if python@2 is installed, which will cause issues with lldb
     if (_commandExistsSync('brew')) {
         const installed = _shell.exec('brew list');
-        if (installed.includes('python@2')) {
+        if (installed && installed.includes('python@2')) {
             console.log('You have Python@2 installed with Brew. Unlinking it since it will cause problems with LLDB');
             _shell.exec('brew unlink python@2');
         }
@@ -272,7 +272,7 @@ async function runOnDevice(selectedDevice, scheme, xcodeProject, configuration, 
         }
     }
 
-    if (pipPackagesOutput.stdout.includes('six==')) {
+    if (pipPackagesOutput.stdout && pipPackagesOutput.stdout.includes('six==')) {
         console.log('Six Python package already installed');
         doInstall();
     } else {
